ff: start forks a watcher process to clean the profile
This changes the behaviour of start to keep running waiting for firefox
to terminate, and then remove the temporary profile.

- While running `start` keeps an instance file in ~/.ff/instances/PORT
  with information about the running instance
- A new command `instances` lists running ff instances, checking the
  contents of the files in .ff/instances

Refactored to stop using the mozrunner FirefoxRunner, implement our own
instead. We need access to the underlying child process to change the
behaviour of start.

`ff start` ignores SIGINT to prevent Ctrl+C from leaving a dangling
session file. Because a shell sends SIGINT to the process group, the
signal reaches the child firefox process and ends ups closing ff as an
accidental consequence.