ff: Reset top frame on each frame iteration
property/attr/text should recusively walk the frames in the page, it
turns out an element ref is relative to the current context. So we
reset back to the top frame on each iteration - this is not correct
but marionette seems to lack references to the current frame, for
now this avoid (part) of the issue.