ff: start option --profile to specify profile path
Allow the user to specify a profile when starting the browser.

Firefox does not allow multiple instances to use the same profile
simultaneously, it produces an error popup when this happens. Internally
firefox enforces this using a lock file stored inside the profile path

    http://kb.mozillazine.org/Profile_in_use

I would like to verify the profile path to check for the lock and avoid
starting ff when we run into this case, however the actual behaviour of
the browser did not match the one described in the link above, so for
now I'm leaving it as it is. I've reduced the number of connection
retries to reduce the wait time before an error is produced.