Add support for async script
Marionette supports async scripts that terminate by calling the
marionetteScriptFinished() with a result value.

For some reason firefox 54 does not accept null as the arguments
attribute in the script passed to executeAsyncScript() so we now
default to an empty array.

ff now has a --async option for the exec command.