Add option to rerun without marionette
Firefox exposes the presence of marionette to remote services via a
javascript variable. This is exploited by multiple services to enforce
checks agains the browser (e.g. cloudflare blocks the client).

Allow a session to be (re-)started without marionette once the initial
setup is complete. In which case there is no remote control.